Function ssg::file::add

pub fn add(path: &Path) -> Result<Vec<File>>

Function: add - returns a Result containing a vector of File structs

Reads all files in a directory specified by the given path and adds them to a vector. Each file is represented as a File struct containing the name and content of the file.

If an error occurs while reading a file, such as the file not existing or being unreadable, an error is printed to the console and the file is skipped. If all files are read successfully, the function returns a Vec<File> containing all the files in the directory.

Arguments

path: A Path struct representing the directory containing the files to be read.

Returns

A Result<Vec<File>, io::Error> containing a vector of File structs representing all files in the directory, or an io::Error if the directory cannot be read.