Function ssg::html::generate_html

pub fn generate_html(
    content: &str,
    title: &str,
    description: &str,
    json_content: Option<&str>
) -> String

Function: generate_html - Generates an HTML page from Markdown

Generates an HTML page from the given Markdown content, title, and description.

This function takes in a Markdown string content, as well as a title and description string to use for the HTML page. The function converts the Markdown content to HTML using the Comrak library, and adds a header and subheader to the HTML page using the title and description strings, respectively.

If content begins with front matter (denoted by “—\n”), the front matter is skipped and only the Markdown content below it is used to generate the HTML. If title or description are empty strings, they are not included in the generated HTML page.

The resulting HTML page is returned as a string.

Examples

use ssg::html::generate_html;

let content = "## Hello, world!\n\nThis is a test.";
let title = "My Page";
let description = "This is a test page";
let html = generate_html(content, title, description, None);
assert_eq!(
    html,
    "<h1 id=\"h1-id-my-page\" id=\"\" class=\"h1-id-my-page\">My Page</h1><p>This is a test page</p><h2><a href=\"#hello-world\" aria-hidden=\"true\" class=\"anchor\" id=\"hello-world\"></a>Hello, world!</h2>\n<p>This is a test.</p>\n"
);