Function ssg::template::render_page

pub fn render_page(
    options: &PageOptions<'_>,
    template_path: &String,
    layout: &String
) -> Result<String, String>

Function: render_page - Render an HTML page

Renders an HTML page with the given title, date, description, keywords, meta tags, CSS file, content, and copyright notice.

This function takes in several arguments that correspond to various parts of an HTML page. These arguments are used to construct a context HashMap that is passed to the render_template function along with a template HTML file. The resulting string returned by render_template is the final HTML page that is generated.

Arguments

author - A string representing the author.

banner - A string representing the banner image.

charset - A string representing the character set.

content - A string representing the content.

copyright - A string representing the copyright notice.

css - A string representing the css.

date - A string representing the date.

description - A string representing the description.

generator - A string representing the generator.

google_site_verification - A string representing the Google site verification.

image - A string representing the image.

keywords - A string representing the keywords.

lang - A string representing the language.

layout - A string representing the layout.

meta - A string representing the meta tags.

bing_site_verification - A string representing the bing_site_verification.

name - A string representing the site name.

navigation - A string representing the navigation.

og_description - A string representing the OpenGraph description.

og_image_alt - A string representing the OpenGraph image alt text.

og_image - A string representing the OpenGraph image.

og_locale - A string representing the OpenGraph locale.

og_site_name - A string representing the OpenGraph site name.

og_title - A string representing the OpenGraph title.

og_type - A string representing the OpenGraph type.

og_url - A string representing the OpenGraph URL.

subtitle - A string representing the subtitle.

title - A string representing the title.

twitter_card - A string representing the Twitter card type.

twitter_creator - A string representing the Twitter creator.

twitter_description - A string representing the Twitter description.

twitter_image_alt - A string representing the Twitter image alt text.

twitter_image - A string representing the Twitter image.

twitter_site - A string representing the Twitter site.

twitter_title - A string representing the Twitter title.

twitter_url - A string representing the Twitter URL.

Returns

If the function succeeds, it returns Ok(html), where html is the HTML page generated by the function. If the function encounters an error, it returns Err(error), where error is a string describing the error that occurred.