Struct ssg::template::PageOptions

pub struct PageOptions<'a> {

    pub author: &'a str,
    pub banner: &'a str,
    pub charset: &'a str,
    pub content: &'a str,
    pub copyright: &'a str,
    pub css: &'a str,
    pub date: &'a str,
    pub description: &'a str,
    pub generator: &'a str,
    pub image: &'a str,
    pub google_site_verification: &'a str,
    pub keywords: &'a str,
    pub lang: &'a str,
    pub layout: &'a str,
    pub meta: &'a str,
    pub bing_site_verification: &'a str,
    pub name: &'a str,
    pub navigation: &'a str,
    pub og_description: &'a str,
    pub og_image_alt: &'a str,
    pub og_image: &'a str,
    pub og_locale: &'a str,
    pub og_site_name: &'a str,
    pub og_title: &'a str,
    pub og_type: &'a str,
    pub og_url: &'a str,
    pub subtitle: &'a str,
    pub title: &'a str,
    pub twitter_card: &'a str,
    pub twitter_creator: &'a str,
    pub twitter_description: &'a str,
    pub twitter_image_alt: &'a str,
    pub twitter_image: &'a str,
    pub twitter_site: &'a str,
    pub twitter_title: &'a str,
    pub twitter_url: &'a str,

}

Struct: PageOptions - Options for rendering a page template

This struct contains the options for rendering a page template. These options are used to construct a context HashMap that is passed to the render_template function.

Arguments

author - A string representing the author.

banner - A string representing the banner image.

charset - A string representing the character set.

content - A string representing the content.

copyright - A string representing the copyright notice.

css - A string representing the css.

date - A string representing the date.

description - A string representing the description.

generator - A string representing the generator.

image - A string representing the image.

keywords - A string representing the keywords.

lang - A string representing the language.

layout - A string representing the layout.

meta - A string representing the meta tags.

msvalidate_01 - A string representing the Bing site verification.

name - A string representing the site name.

navigation - A string representing the navigation.

og_description - A string representing the OpenGraph description.

og_image_alt - A string representing the OpenGraph image alt text.

og_image - A string representing the OpenGraph image.

og_locale - A string representing the OpenGraph locale.

og_site_name - A string representing the OpenGraph site name.

og_title - A string representing the OpenGraph title.

og_type - A string representing the OpenGraph type.

og_url - A string representing the OpenGraph URL.

subtitle - A string representing the subtitle.

title - A string representing the title.

twitter_card - A string representing the Twitter card type.

twitter_creator - A string representing the Twitter creator.

twitter_description - A string representing the Twitter description.

twitter_image_alt - A string representing the Twitter image alt text.

twitter_image - A string representing the Twitter image.

twitter_site - A string representing the Twitter site.

twitter_title - A string representing the Twitter title.

twitter_url - A string representing the Twitter URL.

Fields
author: &'a str

A string representing the author of the page.

banner: &'a str

A string representing the banner of the page.

charset: &'a str

A string representing the charset of the page.

content: &'a str

A string representing the content of the page.

copyright: &'a str

A string representing the copyright notice of the page.

css: &'a str

A string representing the CSS file of the page.

date: &'a str

A string representing the date of the page.

description: &'a str

A string representing the description of the page.

generator: &'a str

A string representing the generator of the page.

image: &'a str

A string representing the image of the page.

google_site_verification: &'a str

A string representing the google site verification of the page.

keywords: &'a str

A string representing the keywords of the page.

lang: &'a str

A string representing the language of the page.

layout: &'a str

A string representing the layout of the page.

meta: &'a str

A string representing the meta tags of the page.

bing_site_verification: &'a str

A string representing the bing_site_verification of the page.

name: &'a str

A string representing the navigation of the page.

navigation: &'a str

A string representing the open graph description of the page.

og_description: &'a str

A string representing the open graph image alt of the page.

og_image_alt: &'a str

A string representing the open graph image of the page.

og_image: &'a str

A string representing the open graph locale of the page.

og_locale: &'a str

A string representing the open graph site name of the page.

og_site_name: &'a str

A string representing the open graph title of the page.

og_title: &'a str

A string representing the open graph type of the page.

og_type: &'a str

A string representing the open graph url of the page.

og_url: &'a str

A string representing the site name of the page.

subtitle: &'a str

A string representing the subtitle of the page.

title: &'a str

A string representing the title of the page.

twitter_card: &'a str

A string representing the twitter card of the page.

twitter_creator: &'a str

A string representing the twitter creator of the page.

twitter_description: &'a str

A string representing the twitter description of the page.

twitter_image_alt: &'a str

A string representing the twitter image alt of the page.

twitter_image: &'a str

A string representing the twitter image of the page.

twitter_site: &'a str

A string representing the twitter site of the page.

twitter_title: &'a str

A string representing the twitter title of the page.

twitter_url: &'a str

A string representing the twitter url of the page.