Function ssg::utilities::minify_html

pub fn minify_html(file_path: &Path) -> Result<String>

Function: minify_html - Minify a single HTML file.

This function takes a reference to a Path object for an HTML file and returns a string containing the minified HTML.

Arguments

file_path - A reference to a Path object for the HTML file.

Returns

A Result containing a string containing the minified HTML.

Ok() if the HTML file was minified successfully.

Err() if the HTML file could not be minified.