Crate ssg

Shokunin 職人 🦀

A Fast and Flexible Static Site Generator written in Rust 🦀

Overview 📖

Shokunin 職人 is a highly-optimized, Rust-based static site generator (ssg) that aims to provide an easy-to-use and powerful tool for building professional static websites and blogs.

The library extracts metadata and content to generate static HTML files from Markdown, YAML, JSON, and TOML. It also supports HTML themes and custom templates to help you create high quality websites with ease.

Features ✨

Blazing fast and flexible

Easy to use

Written in Rust 🦀

Supports multiple content formats (Markdown, YAML, JSON, TOML)

Compatible with various HTML themes and Premium templates to create accessible websites quickly and efficiently

Generates minified HTML and JSON versions for optimal performance

Built-in Rust development server with live reloading

Getting Started 🚀

It takes just a few minutes to get up and running with Shokunin 職人.

Installation

To install Shokunin 職人, you need to have the Rust toolchain installed on your machine. You can install the Rust toolchain by following the instructions on the Rust website.

Once you have the Rust toolchain installed, you can install Shokunin 職人 using the following command:

cargo install ssg

For simplicity, we have given Shokunin 職人 a simple alias ssg which can stand for Shokunin Site Generator or Static Site Generator.

You can then run the help command to see the available options:

ssg --help

Examples and Usage 📚

Check out the examples folder for helpful snippets of code that demonstrate how to use the Shokunin 職人 library. You can also check out the documentation for more information on how to use the library.

License 📜

The project is licensed under the terms of both the MIT license and the Apache License (Version 2.0).

Apache License, Version 2.0

MIT license

Modules

cli

The cli module contains functions for the command-line interface.

compiler

The compiler module contains functions for the compilation process.

file

The file module handles file reading and writing operations.

frontmatter

The frontmatter module extracts the front matter from files.

html

The html module generates the HTML content.

json

The json module generates the JSON content.

metatags

The metatags module generates the meta tags.

navigation

The navigation module generates the navigation menu.

parser

The parser module contains functions for parsing command-line arguments and options.

rss

The rss module generates the RSS content.

serve

The serve module contains functions for the development server.

template

The template module renders the HTML content using the pre-defined template.

utilities

The directory function ensures that a directory exists.

Functions

run

Function: run - Runs the static site generator command-line tool.